Fruit Burahol (Stelechocarpus burahol)
image
image
Nutritious:
On the body odor, the diuretic properties launch urine drainage, traditional pregnancy prevention, blood cleansing, strengthening the liver, lungs, and kidneys. This fruit leaf has a lowering properties of uric acid and lower body cholesterol.

How to consume it:
this fruit skin tastes bitter. Peeling should be done rather thick. This fruit has 8-10 seeds. Sweet and fragrant fruit flesh, but slippery around the seeds. So, be careful when eating this part because the seeds can be swallowed to the throat. this fruit should be consumed directly. Processing of this fruit will actually damage the taste and aroma. The fruit is well consumed each time.